Wire up Rewards Multiplier
