+++ apps/hyperdrive-trading/src/ui/markets/PointsMarkets/PointsMarkets.tsx
@@ -1,25 +1,22 @@
-import { fixed, parseFixed } from "@delvtech/fixed-point-wasm";
+import { fixed } from "@delvtech/fixed-point-wasm";
 import {
   getYieldSource,
   HyperdriveConfig,
-  PointMultiplierReward,
 } from "@delvtech/hyperdrive-appconfig";
 import { Link, useSearch } from "@tanstack/react-router";
 import { ReactElement, ReactNode } from "react";
 import Skeleton from "react-loading-skeleton";
 import { formatRate } from "src/base/formatRate";
-import { calculateMarketYieldMultiplier } from "src/hyperdrive/calculateMarketYieldMultiplier";
 import { useAppConfigForConnectedChain } from "src/ui/appconfig/useAppConfigForConnectedChain";
 import { Well } from "src/ui/base/components/Well/Well";
-import { useCurrentLongPrice } from "src/ui/hyperdrive/longs/hooks/useCurrentLongPrice";
 import { useLpApy } from "src/ui/hyperdrive/lp/hooks/useLpApy";
 import { AssetStack } from "src/ui/markets/AssetStack";
 import { usePoolsList } from "src/ui/markets/hooks/usePoolsList";
+import { usePointsMultipliers } from "src/ui/markets/PointsMarkets/usePointsMultipliers";
 import {
   MARKET_DETAILS_ROUTE,
   POINTS_MARKETS_ROUTE,
 } from "src/ui/markets/routes";
-import { useRewards } from "src/ui/rewards/useRewards";
 import { useYieldSourceRate } from "src/ui/vaults/useYieldSourceRate";
 import { useAccount } from "wagmi";
 
@@ -116,27 +113,7 @@ function PointsMarketCardBanner({
 }: {
   hyperdrive: HyperdriveConfig;
 }) {
-  const { rewards } = useRewards(hyperdrive);
-  const { longPrice } = useCurrentLongPrice({
-    chainId: hyperdrive.chainId,
-    hyperdriveAddress: hyperdrive.address,
-  });
-  const pointRewards = rewards?.filter(
-    ({ type }) => type === "pointMultiplier",
-  ) as PointMultiplierReward[] | undefined;
-
-  const multipliers =
-    longPrice && pointRewards
-      ? pointRewards.map(({ pointMultiplier, pointTokenLabel }) => {
-          const capitalMultiplier = calculateMarketYieldMultiplier(longPrice);
-          return {
-            multiplier: capitalMultiplier
-              .mul(parseFixed(pointMultiplier))
-              .format({ decimals: 0 }),
-            label: pointTokenLabel,
-          };
-        })
-      : [];
+  const multipliers = usePointsMultipliers({ hyperdrive });
   return (
     <div className="flex w-full items-center justify-between rounded-xl bg-base-200 p-5">
       <div className="flex w-full flex-col items-center justify-center gap-1.5">
@@ -144,7 +121,7 @@ function PointsMarketCardBanner({
           Earn up to
         </p>
         <div className="flex w-full justify-between">
-          {multipliers.map(({ multiplier, label }) => (
+          {multipliers?.map(({ multiplier, label }) => (
             <div
               key={label}
               className="flex w-full flex-col items-center justify-center gap-1.5"
@@ -173,6 +150,8 @@ function PointsMarketTable({ hyperdrive }: { hyperdrive: HyperdriveConfig }) {
     chainId: hyperdrive.chainId,
     hyperdriveAddress: hyperdrive.address,
   });
+
+  const multipliers = usePointsMultipliers({ hyperdrive });
   return (
     <div className="flex flex-col gap-3">
       <PointsMarketRow
@@ -264,9 +243,15 @@ function PointsMarketTable({ hyperdrive }: { hyperdrive: HyperdriveConfig }) {
           </span>
         }
         col2={
-          <span className="gradient-text mr-5 text-sm font-medium">64x</span>
+          <span className="gradient-text mr-5 text-sm font-medium">
+            {multipliers?.[0].multiplier}x
+          </span>
         }
-        col3={<span className="mr-3 text-sm">Up to 64x</span>}
+        col3={
+          <span className="mr-3 text-sm">
+            Up to {multipliers?.[0].multiplier}x
+          </span>
+        }
       />
       <PointsMarketRow
         col1={<span className="text-sm text-neutral-content">Term</span>}
